refactor: more comprehensible integer conversion

+++ bridge/jsb_string_name_cache.h
@@ -42,7 +42,7 @@ namespace jsb
             }
             const StringNameID id = values_.add({ p_string_name, {} });
             name_index.insert(p_string_name, id);
-            JSB_LOG(VeryVerbose, "new string name (plain) %s %d [slots:%d]", p_string_name, (uint32_t) id, values_.size());
+            JSB_LOG(VeryVerbose, "new string name (plain) %s %d [slots:%d]", p_string_name, id, values_.size());
             return id;
         }
 
@@ -71,7 +71,7 @@ namespace jsb
 #endif
                 slot.ref_ = TStrongRef(isolate, p_value);
                 value_index_.insert(std::pair(TWeakRef(isolate, p_value), id));
-                JSB_LOG(VeryVerbose, "new string name pair (js) %s %d [slots:%d]", name, (uint32_t) id, values_.size());
+                JSB_LOG(VeryVerbose, "new string name pair (js) %s %d [slots:%d]", name, id, values_.size());
                 return name;
             }
         }
@@ -109,7 +109,7 @@ namespace jsb
                 const v8::Local<v8::String> str_val = impl::Helper::new_string(isolate, p_name);
                 slot.ref_ = TStrongRef(isolate, str_val);
                 value_index_.insert(std::pair(TWeakRef(isolate, str_val), id));
-                JSB_LOG(VeryVerbose, "new string name pair (cpp) %s %d [slots:%d]", p_name, (uint32_t) id, values_.size());
+                JSB_LOG(VeryVerbose, "new string name pair (cpp) %s %d [slots:%d]", p_name, id, values_.size());
                 return str_val;
             }
             return slot.ref_.object_.Get(isolate);
